waf/requirements.txt
aiohttp==3.9.1
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.28.0
python-dotenv==1.0.0
requests==2.31.0
mitmproxy==10.1.5
psycopg2-binary==2.9.9
setuptools==69.0.3
